adding rest api

+++ Online-Shop/src/main/java/com/example/OnlineShop/Database/Dtos/ProductLessDto.java
@@ -2,7 +2,7 @@ package com.example.OnlineShop.Database.Dtos;
 
 public record ProductLessDto(	Long id,
 								String name,
-								float price,
+								double price,
 								String seller) {
